create a dedicated sample connector class for tests
