+++ connect/runtime/src/test/java/org/apache/kafka/connect/runtime/AbstractHerderTest.java
@@ -1335,10 +1335,10 @@ public class AbstractHerderTest {
 
     @Test
     public void testDifferenceInConfigDefAndValidation() {
-        AbstractHerder herder = createConfigValidationHerder(SampleSourceConnector.class, noneConnectorClientConfigOverridePolicy);
+        AbstractHerder herder = createConfigValidationHerder(SampleImbalancedConfigVerificationConnector.class, noneConnectorClientConfigOverridePolicy);
 
         Map<String, String> config = new HashMap<>();
-        config.put(ConnectorConfig.CONNECTOR_CLASS_CONFIG, SampleSourceConnector.class.getName());
+        config.put(ConnectorConfig.CONNECTOR_CLASS_CONFIG, SampleImbalancedConfigVerificationConnector.class.getName());
         config.put("name", "somename");
         config.put("required", "value");
         config.put("testKey", null);
